requirements.txt
Django==2.0.5
django-allauth==0.35.0
django-environ=0.4.4
